Dictionary and Translator
The translator file reads in words in english and elvish from the dictionary, then allows you to type sentences and receive the words translated in elvish. In the event elvish words are not present the english word is printed instead. Currently, this does not account for grammar, punctuation, or other such edge cases and has a rather limited dictionary. I intent to update the dictionary with new words as I create them and hope to add some logic for grammar and punctuation.